add shm_open/shm_unlink to fslib [closes #2265]
Its not entirely fixed actually.
The group used is still "dosemu2", only the owner is set to
current user. This is to be addressed later.

Also the G and O Read perms are added because the reading is done
on behalf of a new user, and this probably can't be changed as
the read is being done from dlopen() which doesn't have an fd
variant. dlopen() from /proc/self/fd/xx will likely not change the
perms, as /proc/self/fd/xx is just a symlink.