init: alloc dpmi mem from shm [#2294]
This increases RssShmem from 16Mb to ~900Mb. :(
Seems RssShmem accounts aliases after CoW. I.e. initially it shows
~700Mb (512Mb dpmi_base + 128Mb dpmi_mem + 16Mb xms + 32Mb align_wnd).
But as writing happens, RssShmem increases, presumably because aliases
are getting CoWed on VMA level (but not on a PTE level!) and linux can't
track the aliasing any more.
That sounds like a broken/inaccurate accounting.
But 700Mb is also quite bad. :(