simx86: fix off-by-1 page invalidation logic
Invalidation length was wrongly calculated: if the last byte is
byte 0 on the page, then this page wasn't taken into account.
Eg the len of 1 was becoming PAGE_SIZE if the address is not
page-aligned, but if it is page-aligned, then the len would be 0!

This fixes some of the Jazz Jackrabbit problems, namely
"touched JIT-protected page" errors. There are also gfx glitches
that needs to be debugged, as well as occasional slowdowns and
other oddities.