dnative: fix race with stalled errcode
Async signals may contain stalled errcodes, including the ones
representint INT instruction. If we happen to hit SIGALRM on HLT
or HLT is set by RSP or LDT call, then the speculative hlt handler
was raising INT based on a stalled errcode value.

This patch zeroes out errcode in async signals, and also adds a
check to speculative hlt handler that makes sure INT can be raised
only when the appropriate opcode is seen.

This fixes a rare crash of DPMI apps on mouse move.
Mouse move involves lots of RSP calls for client switch, if comcom
is started with /M2 switch, which is the default.
That was triggering the race.