mapping: fix horrible overcommit bug
All allocated mapping was pre-committed!
This was not a problem before, but I obviously overlooked that
part when made full memory alias. As the result, dosemu2 committed
1.5Gb on start, OOMing the 1Gb qemu box immediately.

With this patch we only pre-commit whatever smalloc allocates.
Everything else gets committed on demand.

Deep apoligize to all the Raspberry-PI users.
No one complained though. :)