dlmalloc: use MAP_32BIT flag [fixes #2317]
We use dlmalloc only for JITted code.
MAP_32BIT is needed because the cross-node jumps (inserted by linker)
are only possible within the +/-2Gb range. When the range is larger,
the jump was happening to unmapped location.

That should fix all the weird linker instabilities seen here for
years, but they were never reproducible until now.