simx86: pass full cr2 to e_vgaemu_fault()
Page number was passed instead (cr2 right-shifted to PAGE_SHIFT),
which was shifted back before checking things like memcheck_is_rom().
But not everything is page-aligned in DOS memory map, so this
double-shifting was able to hit different area and mistakenly
confuse UMB with ROM.

This became especially visible with 16Kb pages.