jit: get rid of the omit-frame-pointer games
Just preserve $bp by hands instead of trying to clobber it.
Its quite difficult to use push for that, as this modifies $rsp so
the mem-ref via rsp breaks. And using "r" is no go on i686, where
we get out of regs. The "clever trick" is to use xchg with "m".
This way we can manually preserve $ebp.